Fix comparing lodOrBias as integer values
lodOrBias can be either a floating-point LOD, a floating-point bias,
or an integer LOD. The latter is used by OpImageFetch SPIR-V
instructions (texelFetch in GLSL). To check if they diverge, we have
to use integer comparison, since floating-point comparison may treat
them as subnormal values which compare equal to zero when DAZ is
enabled.

Bug: b/169904252
Test: dEQP-VK.texture.subgroup_lod.texelfetch
Change-Id: Ifaceed84b09f8e3f1d31983cf14883343aeac72f