Clamp depth before the depth test
Previously we clamped fragment depth right after interpolation, and
after shader execution for shader-provided depth. This means logically
it could be clamped twice, although only one of them would actually be
used. It was also confusing to understand where it was subsequently
consumed, and didn't closely follow the Vulkan spec which considers
depth clamp a part of the depth test (but, importantly, takes place
regardless of the value of depthTestEnable).

This change performs depth clamping right before the actual depth test.

Bug: b/185227903
Change-Id: I4307b9beef2e1a10606bf9accac5c6b86daa6d67