Skip depth bounds test when there is no depth attachment
The Vulkan spec states that "If the depth bounds test is disabled, or if
there is no depth attachment, the coverage mask is unmodified by this
operation."

Previously the depth bounds test was performed even when there was no
depth attachment. We were also not obtaining the depth buffer format
unless the regular depth test was enabled.

This was discovered while refactoring to assert that the depth buffer
had one of the expected formats, but it was VK_FORMAT_UNDEFINED in tests
like e.g. dEQP-VK.dynamic_state.ds_state.depth_bounds_1.

Bug: b/181656417
Bug: b/185227903
Change-Id: Ic0a06b668c85eee7565a9007f7efb77621f3f7ac