Suppress deprecated iterator warning in GN build of LLVM 10.0
std::iterator<> is deprecated in C++17, which produces a warning in
recent versions of Clang, which we treat as an error. This change
suppresses the warning by using -Wno-deprecated-declarations.

While this flag was already used for GN builds of our LLVM 10.0 copy,
Reactor includes headers such as llvm/include/llvm/IR/Value.h which
also use std::iterator<>.

Note that Subzero appears to be affected by the same issue due to its
use of llvm::Value, which this change doesn't address yet.

The root cause of using a deprecated feature was fixed in upstream
LLVM at https://reviews.llvm.org/D67586

Bug: b/198636514
Change-Id: I473e153ca7243a824a0f6835094381448c0b61d1