Compute the image size in 64-bit arithmetic
The size of a 2D 'slice' is returned in a 32-bit value, which is
sufficient for the maximum supported dimensions of 8192x8192, but for
3D images we need to multiply by the depth in 64-bit to avoid numerical
overflow.

Note that currently we only allow memory allocations up to 1 GiB, but
this change fixes correctly reporting how much memory would be required,
from a call to vkGetImageMemoryRequirements().

Bug: chromium:1248567
Change-Id: Ic055790b4ae53355a90a173b824e94e3c5933316