Regres: Skip the Vulkan Loader
Previously we relied on setting VK_ICD_FILENAMES to point the system's
Vulkan Loader library (libvulkan.so.1) to SwiftShader's ICD. This change
makes Regres directly load SwiftShader's build of the libvulkan.so.1
library, which exports all of the API entry functions too. Loading of
Vulkan Layers is not supported by it.

This has the advantage of not depending on the version of the Vulkan
Loader, which can affect Regres' test results. The system running
Regres doesn't even require having a Vulkan Loader at all this way.
It is also measurably faster.

The compromise is that it currently causes 1 test to fail:
dEQP-VK.api.version_check.entry_points, due to the lack of support for
loading layers. Note that implicitly loaded layers could also affect the
Regres results, so this isn't necessarily a bad thing.

We also wouldn't be alerted if changes broke the compatibility of our
ICD with the Vulkan Loader. This is unlikely to happen though if tests
that are being run during development (e.g. using Cherry) are set up to
use the Vulkan Loader, especially when touching code that affects this
interface. In addition, downstream tests (e.g. Chromium) would quickly
detect any bugs that slip through.

Note that this change works by appending the LD_LIBRARY_PATH with the
path to SwiftShader's libvulkan.so.1, not prepending it, so one can
still point LD_LIBRARY_PATH to a Vulkan Loader library and not have it
being skipped.

Bug: b/200679975
Change-Id: I0ba2dd06fa1ca0749e852b1dc62955199f61c6b3