Fix obtaining the execution model from the intended enty point
Previously the `executionModel` member was overwritten for each
OpEntryPoint that was being parsed.

This change also refactors constant members that are initialized in the
constructor to not have an in-class initialization value, since those
can be confused to only have that value (like constexpr) or can lead to
forgetting to initialize them in the constructor as intended.

Bug: swiftshader:161
Change-Id: Ie0792bff3e68bbf4bfa2873a189fa6811e508e60