Don't initialize memory allocations
Vulkan does not make any guarantees about allocations being zero-
initialized, so applications have the responsibility of clearing it
themselves if they need it to be initialized.

This change avoids redundant initialization, speeding up startup time
and avoiding jank.

Note that this change only affects sw::allocate(), which is used by
vkMemoryAllocate() for Vulkan resources, and by the graphics and compute
pipelines for intermediate storage. The former is under the
application's control, while the latter is SwiftShader's responsibility
to initialize where necessary, which is extensively tested by
MemorySanitizer builds.

Reactor's allocation of executable memory for JIT-compilation purposes
is specifically still zero-initialized (except in MemorySantizer builds

Bug: b/140991626
Change-Id: I73d1f62536b8812997546973faf0ca3cdd55679e