Deep copy specialization info
Previously a shared_ptr<> was used to avoid copying specialization info,
but it had the downside of requiring an additional allocation for the
VkSpecializationInfo structure itself, and treated a null pointer
differently from a structure with zero entries.

We copy the specialization info three times: to create the pipeline
cache's lookup key, to copy it into the cache when a new entry is
created, and when running spvtools optimization passes that apply the
specialization constants. Only the second one was a shallow copy, so
this change which makes it a deep copy while avoiding additional
allocations and pointer indirections most probably has no measurable
negative effect.

This change also replaces the use of vk::allocate() with
sw::allocateUninitialized(). We immediately memcpy() the same amount of
data, so it doesn't stay uninitialized. Also note that while this does
not make proper use of allocator callbacks, the old code didn't either,
and the use of the functions in the sw namespace make it clearer that
this is an unaccounted internal allocation.

Bug: b/140991626
Bug: b/172839674
Change-Id: I5962e028bbb6ad44b6dc2874eae6636bca0115c3