Regres fast forwards branch if no SHA is specified
Normally you must provide a commit hash in deqp.json. However, we want
to start testing against tip-of-tree dEQP. So specifying just a branch
makes regres use the tip of that branch when it first checks out dEQP.

This branch will not remain up-to-date for daily runs, but it's the
first step in enabling such behavior.

Bug: b/201453797
Change-Id: I31389a1723b9e6f66edd45529c8fb64bbd24e86c