Don't initialize device memory allocations in MSan builds
This allows detecting more potential uses of uninitialized memory in
MemorySanitizer (MSan) builds.

It is the application's responsibility to clear memory allocated by
vkAllocateMemory(), or guarantee that it will be written to before being
read.

Non-MSan builds are unaffected by this change, so that any errors that
are found in the near term don't pose a security risk. A subsequent
change will remove initialization for non-MSan builds as well.

Bug: b/140991626
Change-Id: I079776345c84567875e65e3c390d2b0a29e507db