Refactor DeviceMemory buffer allocation/freeing
This change completes replacing "deallocate" with "free" to match the
Vulkan nomenclature.

DeviceMemory::size was renamed to allocationSize to match the
VkMemoryAllocateInfo field and avoid potential confusion with other
sizes. The virtual DeviceMemory::allocate() method was only ever called
with the allocation size and a pointer to the (previously private)
buffer member as parameters. So it was renamed to allocateBuffer() and
doesn't take any parameters so each implementation sets the member
variable directly. Likewise deallocate() previously took the same
parameters and has been replaced by freeBuffer().

Bug: b/134584057
Change-Id: I401b4dad2632593dd7e155480475f0f7f20788f6