Avoid recompiling identical SPIR-V binaries
We previously took the serial ID from the unoptimized SPIR-V module, for
use as part of the Routine cache keys. If a pipeline is created with
specialization constants, we assumed them to produce unique optimized
SPIR-V binaries, regardless of their actual values, and obtained a new
serial ID. If a shader module was used with identical specialization
data, we would redundantly JIT-compile the same routine if no other
state used by the pipeline stage was different.

This change takes advantage of the fact that the pipeline cache doesn't
add a new entry (of optimized SPIR-V) for an already optimized module
unless the values of the specialization constants are different and
thus would produce a distinct binary (assuming said specialization
constants affect the code). A new serial ID is associated with every
new SPIR-V binary. For a (pipeline) cache hit, the existing entry's
ID is used as part of the lookup key for the routine cache.

Note an application has to use a VkPipelineCache object to take
advantage of this.

Bug: b/172839674
Change-Id: I18b937800ea021235081bba0d827436265816414