Fix border update when rendering to cube face
Any notification that the contents of an image has been modified was
marked as requiring the image to have the VK_IMAGE_USAGE_STORAGE_BIT
set. This was only true for a subset of the cases and most cases
actually perform modifications using a direct memory accesses on the
vk::Image memory, so this requirement was wrong most of the time.

This CL adds an explicit argument to ImageView::contentsChanged() so
that we can specify if the VK_IMAGE_USAGE_STORAGE_BIT is expected to
be set or not, depending on where the update comes from.

Bug: b/197658470
Tests: dEQP-VK.image.sample_cubemap.write_face_0
Change-Id: I8290e329fa13215c23b31feef4fc8de391ce8032