Allow compressed images to blit
The Vulkan spec allows blitting the result of image decompression
into a format compatible with the decompressed format. This CL
allows Image::blitTo() to perform this task.

As a side effect, since Image::blitTo() is a const function called
from a const object, which is expected, all Image decompression
functions are now marked as const and only the dirtySubresources
member along with its associated mutex are marked as mutable.

Tests: dEQP-VK.api.copy_and_blit.*
Bug: b/202978657
Change-Id: I46e0403c3ec13c06b0c7456a29ee16c7bc7628ca