Avoid SPIR-V binary identifier clashes
If we ever use both unoptimized and optimized SPIR-V binaries (e.g. as
part of tiered JIT compilation), we need them to have different
identifiers. This is accomplished by using the bitwise inverse of the
unoptimized binary's identifier as the one for the optimized binary.

Note this only affects cases where no pipeline cache is being used. If
one is used, we use the identifier that got assigned when creating the
cache entry.

Bug: b/172839674
Change-Id: I894a01f21518b046845cd56dcbb0d13cea1ff94b