Fix uninitialized memoryOwner on unused DescriptorSet
In the Renderer, we call DescriptorSet::PrepareForSampling() on all
input descriptors, not only the ones used in the current draw pass.
Because of this, DescriptorSet::PrepareForSampling() might be called
on uninitialized descriptors. Before this CL:
https://swiftshader-review.googlesource.com/c/SwiftShader/+/57889
that memory used to be zeroed out, so that wasn't an issue, but now
the memoryOwner pointer may be used without being initialized. This
CL adds an initialization for the memoryOwner field on structures
which use it (SampledImageDescriptor and StorageImageDescriptor).

Bug: chromium:1260606
Change-Id: I41a02a8b652a48b790d8aacf16fd28d26e527336