Refactor Queue::Submit to use internal structure
Before this CL, Queue::submit() would internally use the VkSubmitInfo
structure to store the queue submit related information. This had the
downside of having to re-parse the pNext chain in submitQueue(),
which is now no longer the case in this CL, since the internal
structure no longer includes a pNext pointer.

Also, pointers in the internal structure with a size of 0 are no
longer assigned a value and are set to nullptr instead.

This CL paves the way for the VK_KHR_synchronization2 implementation,
which will be able to use the same internal structure as the one used
by Queue::submit().

Bug: b/204110005
Change-Id: Ib99261ef101680519bed57892a284ae20b229661