Refactor blending
This change mainly renames `oC`, which was the name for a Direct3D 9
color output register, to something that is more readable.

Also the `one` variable in `alphaBlend` was eliminated, since there are
no unnormalized formats which are blendable.

Also we no longer discern between VK_FORMAT_R16_SINT and R16_UINT in
writeColor, since the code paths only differed in casting Int to Short
or UShort, which are identical truncation operations.

Bug: b/134584057
Change-Id: I855d40746b405fb7873cf2a800ec26c6ad533daf