Add configurable subpixel bits rasterization extension.
This allows ANGLE to specify a required subpixel precision in the
rasterization state of the pipeline create info. This in turn will
allow ANGLE to render several GLES traces correctly with overlapping
geometry. We specify this as a tiny pipeline rasterization create
info structure with the bits in a member field.

Bug: b/204560787
Change-Id: Ica1f44bd69f1bc09b02cfc136af6c852530e9f56