Tightly pack PlaneEquations array
interpolateAt* functions receive their parameters at runtime rather than
at compile time, so they require information about how interpolants
are computed inside the setup routine. Because plane equations were
not tightly packed, there needed to be a distinction between arrays
and non arrays, as they were addressed differently. This check was
failing in certain cases and the information to determine whether an
interpolant was within an array wasn't available at this point.

In order to solve this issue, this CL tightly packs the plane
equations array, which allows us to only use the trivial path in
SpirvShader::Interpolate() and remove the need for a distinction
between the array and non-array cases.

Tests: dEQP-VK.pipeline.multisample_interpolation.*
Bug: b/179925303
Bug: b/189588731
Bug: b/183693496
Bug: b/205237948
Change-Id: I87be2a8ccf14261e6ad33a6383e605cfbfea5842