Check for valid SPIR-V <id>s by comparing against 0
This change consistently uses if(id != 0) { ... } syntax for choosing
code paths where the SPIR-V <id> is valid. This allows the value()
method to assert that we don't inadvertently use an <id> for an operand
which does not exist. This makes for an elegant syntax when parsing
SPIR-V instructions into structures which store <id>s for operands which
are optional, such as common in image instructions, while catching
mistakes centrally.

Bug: b/205642050
Change-Id: I3e1f1340a56bd3c848fdf3724088660046bf29ed