Check for valid SPIR-V <id>s using Boolean evaluation
This change allows using if(id) { ... } syntax for choosing code paths
based on whether a SPIR-V <id> is valid. This will allow simplifying the
parsing of SPIR-V instructions by storing <id>s for operands which are
optional, such as common in image instructions.

An assert was added to ensure we don't inadvertently use an <id> for an
operand which does not exist.

Bug: b/205642050
Change-Id: I3e1f1340a56bd3c848fdf3724088660046bf29ed