Simplify obtaining the type from a SPIR-V object <id>
This change replaces nested getType(getObject(id)) calls with a single
getObjectType(id).

Note that we can't use a getType() method which takes an Object::ID,
because we often call it with a typeless uint32_t obtained directly from
the SPIR-V binary with the word() method. We would essentially have to
call getType(Object::ID(id)), so getObjectType(id) is more compact.

Also, redundant casts of resultTypeId() to Type::ID have been removed.

Bug: b/203730083
Change-Id: I4066dc08022f5431c1fdbad25b5f02fa0764085e