Separate the SPIR-V image and sampler as part of ImageInstruction
Previously ImageInstruction::sampledImageId represented either a SPIR-V
Sampled Image object or an Image object, based on the instruction
opcode. Also a Sampled Image object could either represent a combined
image/sampler descriptor or use an OpSampledImage instruction to have
separate image and sampler descriptors.

This change replaces it with two separate members that represent an
image and a sampler object. For combined image/sampler descriptors they
have an identical value, and for samplerless instructions the samplerId
member remains 0. This results in less ambiguous code and avoids SPIR-V
parsing logic in the functions that emit Reactor code.

Bug: b/203730083
Change-Id: I9de0e5067cd76b3159c50c3cdc359d864e9033bf