Determine SPIR-V zero constants from SPIR-V binary only
Previously the Operand::isConstantZero() method would tell us whether
an operand passed to an instruction is actually a constant expression
with a value of zero. However, the Operand class is a wrapper for
Reactor variables, which also supports constants, so we need to be
actively emitting the Reactor routine that compiles the SPIR-V code, to
use it.

Since determining whether an object represents a value of zero can be
done entirely from the SPIR-V code alone, this change moves this method
to the Object class.

Concretely it has the advantage of being able to tell whether an image
instruction has a Sample operand which is zero, so it ca be ignored,
without passing around the EmitState pointer.

Bug: b/203730083
Change-Id: Ie8e6eb819802d5f9662c727ef81d04f47a9610ad