Implement shaderStorageImageWriteWithoutFormat support
When the OpImageWrite instruction is provided with an 'Unknown' Image
Format, use a JIT trampoline to generate a function specialized for the
format provided though the image's descriptor at shader execution time.

This reuses the sampling routine caching mechanisms used by
OpImageSample* instructions.

When the SPIR-V code does provide an Image Format, we 'inline' the
OpImageWrite instruction, emitting the operations as part of the
generated shader routine.

Bug: b/203730083
Tests: dEQP-VK.*without_format*
Change-Id: I3d40059905f772493b2fdf3134ebe0a11a5bb5bb