Provide StorageImageDescriptor with an image view identifier
This will enable OpImageWrite to obtain the image view format through
the identifier, to support the 'Unknown' SPIR-V Image Format.

This change derives both StorageImageDescriptor and
SampledImageDescriptor from a common ImageDescriptor structure which
contains the image view identifier. This way we don't have to worry
about which descriptor type we're dealing with when accessing the
identifier.

Bug: b/203730083
Change-Id: I5d21ee8a576aa54bbc60261818b95e6d050541bc