Perform image write from a static function
The static WriteImage() function generates all the code for an
OpImageWrite instruction without depending on any SpirvShader
data.

This will make it suitable for producing a standalone function similar
to an ImageSampler function, with all the state coming from
ImageInstructionSignature and the StorageImageDescriptor that gets
accessed at shader execution time.

Bug: b/203730083
Change-Id: Ic8c018bca1490a97d833c4cb5be0a6a7338c048b