Support writing to storage images with packed formats
shaderStorageImageWriteWithoutFormat will allow OpImageWrite to write to
images with packed formats.

Note we already claimed VK_FORMAT_FEATURE_STORAGE_IMAGE_BIT and
VK_FORMAT_FEATURE_STORAGE_TEXEL_BUFFER_BIT support for these formats,
but without shaderStorageImageWriteWithoutFormat or
shaderStorageImageReadWithoutFormat they could neither be read nor
written since the spec states that if "The SPIR-V Image Format is not
compatible with the image view’s format", "the value of the texel
returned is undefined." and compatibility is defined as a strict 1:1
mapping.

Bug: b/203730083
Change-Id: I530780f955caefc8a7324f2ab8e05a503f2c20ef