Implement Assert() for Reactor
Assert() is a macro which inserts an assertion into Reactor code.
It follows the C++ requirements: https://en.cppreference.com/w/cpp/error/assert

One must include Reactor/Assert.hpp to make use of it.

The 'death' unit test fails on Kokoro's macOS bots with "died but not
with expected error", but passes locally. This is assumed to be caused
by stderr redirection and should not interfere with the ability to catch
errors using Assert().

Bug: b/156389924
Change-Id: I96f19aa907e6d8cee315a374441bb7a9554b4441