Prevent exit-time destructors in src/WSI, Pipeline, and System
This change adds -Wexit-time-destructors to the Clang compile options
for the CMake build of thesrc/WSI, src/Pipeline, and src/System
subprojects. It causes the Clang compiler to issue an error when an
object with static lifetimewould require executing a destructor at
'exit' time. For SwiftShader, that is when the shared library is
unloaded.

The src/Vulkan and src/Reactor directories still contain source files
that require exit-time destructors.

Bug: b/208256248
Bug: swiftshader:118
Change-Id: I90090ef6d8cf83c60fa3bdde2c6ff85cc5e6f52a