Eliminate SpirvShaderInstructions.inl
SpirvShaderInstructions.inl has caused some issues on a few occasions,
due to requiring to remain in sync with the spv::Op enum. It mostly
matches spv::HasResultAndType(), which is part of the official spirv.hpp
header and thus is kept in sync with the opcode enum upstream. Thus this
change replaces it with that, but instructions which previously were
considered to have side-effects have been preserved.

Bug: b/178477034
Change-Id: I19eb757a9f244872afe59f343b99674473e52e4d