Regres queries gerrit with a CL's numeric ID
Gerrit's API cannot resolve a query for a commit ID if there are
multiple CLs on different branches that have identical commit IDs.
However, numeric IDs are created by gerrit and are unique for each
change within a project. So we can safely use these numeric IDs to
specify a CL.

Bug: b/207382281
Change-Id: I1c80d9d80e67179552c8c8cdc89f97485450583f