Simplify scaling of 16-bit sampled texel components
For image formats that get sampled using v4i16 vectors as intermediate
results we place each texel component's bits in the upper part of the
16-bit vector components. On conversion to floating-point we need to
(un)scale. For example, a 565 format uses 0xF800 to represent 1.0 in the
16-bit vector components so this is the scaling factor used for
normalization.

Previously, determining the scaling factor used the Format::getScale()
method, which returns the scale for when the texel's fields are placed
in the lower part of integer vectors. This floating-point value was
converted to integer, shifted, and then converted back to floating-point
to get the scaling factor for fields placed in the upper bits instead.

This change simplifies it by computing the value which represents 1.0
using just integer arithmetic. For signed components the sign bit is
masked off. The max() operation has been removed since all formats which
are sampled using vectors with 16-bit components as intermediate result
must have no more than 16 bits per component field.

Bug: b/204709464
Change-Id: Id9365065f3d171f67ba2bcf6de5329d323eaf2a9