Report storage usage capability for surface images
VK_IMAGE_USAGE_STORAGE_BIT is practically universally supported for
swapchain surfaces. It enables compute shaders to write directly to
them.

This usage capability is always enabled on Android, which provides a
swapchain surface implementation based on the Android Hardware Buffer
extension, instead of having the ICD provide it. Thus enabling this
capability on other platforms ensures we test the mandatory underlying
support with dEQP-VK.

Bug: b/170414896
Change-Id: I15284b4c6f7105f95dce6d621876399066d24944