Provide a smoother window resize experience
When using Chromium with SwANGLE as the GPU, Window resizes
would cause the image to stutter during resize and
sometimes even fail to update at the end of the resize
operation. To provide a better resize experience, this CL
change the logic on all platforms so that, when returning
VK_ERROR_OUT_OF_DATE_KHR, the presentation is not skipped,
providing a smoother feel on resize operations.

Bug: swiftshader:166
Change-Id: I2bb1d395781c5c8b1e9dfcd86c060f01e69185af