Optimize mirror addressing mode
The previous implementation required a total of 8 instructions. The new
version only takes 6.

Note that the previously used Frac() function includes a Min() operation
which ensures that it never returns results greater or equal to 1.0. It
also handles Inf values. The new mirror addressing implementation has an
explicit Min() which serves similar purposes.

Bug: b/194635449
Change-Id: I3f16beb0ecd93e5bafccacd7b5823710c6ace55e