Make image descriptor fields unsigned
All Vulkan width/height/depth variables are unsigned, while our
descriptor and sampling code used signed integers in various places.
While this was benign it added confusion about whether some of these
values could be negative and need special handling.

Note that the results produced by the 32-bit SamplerCore::address()
method are signed, because negative values represent out-of-bounds
sampling. So these are left unchanged and the image dimensions are
explicitly cast to signed where necessary.

The Mipmap::onePitchP field is also still signed, because it is used
in MulAdd() intrinsics which multiply pairs of signed 16-bit fields
and produce a 32-bit signed result (there is no unsigned PMADDWD
instruction variant). Note that multiplications which produce a result
of the same bit width as the operands are indifferent to signedness
(just like addition and subtraction).

Some assert_cast<>'s were used to catch unexpected overflows, and uses
of `auto` in the descriptor writing code were replaced with explicit
types to aid readability.

Bug: b/210880838
Bug: b/200806413
Change-Id: I8d8cb1b17cfd2d40b0ad3153e75c69a086e2ae40