Expose VK_GOOGLE_decorate_string
SPV_GOOGLE_decorate_string was incorporated into SPIR-V 1.4, which we
already supported. The string decorations are ignored since they're
assumed not to have semantic value for the driver.

OpDecorateStringGOOGLE is numerically equal to OpDecorateString and
OpMemberDecorateStringGOOGLE is equal to OpMemberDecorateString. The
only supported string literal decoration is UserSemantic, so an assert
was added.

Bug: b/214576937
Change-Id: I0fe8d4da60473341c80cc7e8be2f302a79135613