Implement VK_GOOGLE_user_type
"This extension provides one new decoration, UserTypeGOOGLE, that
allows HLL shader compilers to optionally embed unambiguous type
information.

The decoration UserTypeGOOGLE is used with OpDecorateStringGOOGLE and
OpMemberDecorateStringGOOGLE to store a string name for the type of the
decorated object in the user’s source. It can decorate only a variable,
a type, or a member of a structure type. This has no semantic impact."

Bug: b/214576937
Change-Id: I063f10774072c172d627f400f3300e230c5f92b9