Optimize Abs() using LLVM intrinsics
The `fabs` abstract LLVM intrinsic translates to an x86 `vandps`
instruction with a broadcasted scalar memory operand on CPUs that
support EVEX encoding (AVX-512).

The x86::pabsd() intrinsic was removed because it is currently unused
and its use should be discouraged. The integer Abs() operation already
translates into a `vpabsd` instruction.

Bug: b/215191437
Change-Id: Ibc6016ec7d2eae0d6c1cd339b47169defa3fa1eb