Add benchmarks for trigonometry and transcendental functions
Note that the math functions in the rr namespace are 'full' precision as
specified by the standard C/C++ math library (e.g. libm), and are
intended to stay that way. SwiftShader should use more optimal
implementations that satisfy the Vulkan precision requirements, but
keeping these functions in Reactor and benchmarking them is useful as
a baseline, reference, and fallback.

Bug: b/169754022
Change-Id: Ie2f62c7c7b72c252d3ba0733300501aca746609b