Optimize SPIR-V sin and cos operations
This change replaces the previous implementation of sine and cosine
which took 18 multiplications and a division, with one which one takes
5 multiplications.

The derivation of the algorithm is documented in a PDF, and an
exhaustive unit test proves it is accurate to 2^-12 absolute error,
which exceeds Vulkan's 2^-11 tolerance. We also achieved exact results
for x=0, x=pi/6, x=pi/3, and x=pi/2.

Bug: b/169754022
Change-Id: I9e6192f24ebc6e919713e8c197a21215ec562495