Add a simple configuration file implementation.
This is very loosely based on the legacy SwiftConfig, but
it does not come with a web server-based alternative. The
configuration is read from SwiftShader.ini, if the file exists.
This simplifies things and allows us to keep a single Configuration
object around, as the configuration is static. A more complex
solution can be added later when deemed to be necessary.

For now, the only option supported is ThreadCount, which is
used when creating the marl scheduler. To add new options in the
future, one can simply update the Configuration struct and
the parsing code in SwiftConfig.cpp.

Change-Id: I8458a99fc4c3e2d180bbc391b0e12d789dea5dcf