Add AVX2 and FMA support detection
The AVX2 feature bit corresponds to bit 5 of the EBX register after
issuing the CPUID instruction with EAX=7,ECX=0. This extended CPUID
functionality is in turn supported if AVX is supported, which is
indicated by bit 28 of ECX for CPUID issues with EAX=1.

We also check for OSXSAVE and FMA support as part of AVX2.

[0] https://en.wikipedia.org/wiki/CPUID#EAX=1:_Processor_Info_and_Feature_Bits
[1] https://en.wikipedia.org/wiki/CPUID#EAX=7,_ECX=0:_Extended_Features

Bug: b/214591655
Change-Id: Ia16d8d99bd8be7e33d6f823475b5b3951b5d23f9