Implement FMA() which always returns a fused multiply-add result
This Reactor intrinsic falls back to calling std::fma() when the backend
or the CPU does not support FMA instructions. The Caps::fmaIsFast()
function can be used to check whether FMA instructions will be emitted
instead.

Note that we rely on LLVM to emit FMA instructions when AVX2 with FMA
support is detected on x86 processors. We test this assumption for the
rr::MulAdd() intrinsic.

Bug: b/214591655
Change-Id: Ide069606cd9fe7cc40fa42e68719b24a0046405d