Implement FMA() which always returns a fused multiply-add result
This Reactor intrinsic falls back to calling std::fma() when the backend
or the CPU does not support FMA instructions. The rr::Caps::fmaIsFast()
function can be used to check whether FMA instructions are highly likely
to be emitted instead.

This intrinsic must be used for algorithms that demand the precision of
FMA operations. rr::MulAdd() cannot be relied upon to produce FMA
results, even if fmaIsFast() is true.

Bug: b/214591655
Change-Id: Ide069606cd9fe7cc40fa42e68719b24a0046405d