Add a query for fast FMA support
If rr::Caps::fmaIsFast() returns true, the backend and the CPU are known
to support FMA instructions and they are likely to be used by the
rr::MulAdd() intrinsic. If false, it is unknown whether it behaves as a
multiplication followed by an addition, or uses FMA (in other words,
this function is conservative).

Currently only the LLVM backend supports FMA instructions, and detecting
CPU support is only implemented for x86. We assume that LLVM will emit
FMA instructions for its `fmuladd` intrinsic when AVX2 is supported. A
test was added to verify this assumption.

Still, fmaIsFast() must not be relied upon for selecting between
algorithms that demand FMA support, and ones that don't, when only using
the rr::AddMul() intrinsic. Backends may still implement them using
multiplication followed by addition even if FMA instructions are
supported. A future change will add rr::FMA() which is guaranteed to
either emit FMA instructions or call std::fma(). fmaIsFast() can then be
used to select between an algorithm which uses FMA(), and one that does
not, without risking correctness issues.

Bug: b/214591655
Change-Id: I33bfaa788349b567c2245dd018e96f7868e3a9a5