Copy Reactor transcendentals benchmarks to new PipelineBenchmarks
Benchmarking Reactor's transcendental functions as part of
ReactorBenchmarks was not that useful because they're just meant to be
reference implementations and we can't compare them to other
implementations external to Reactor.

The new PipelineBenchmarks target gives us access to the optimized
'ShaderCore' implementations of SPIR-V GLSL.std.450 compliant
transcendental instructions. For now just benchmark the Reactor
intrinsics as a baseline.

Bug: b/158231104
Change-Id: I838ec135ad35732c0173a6866eab07eb5fcfe1b2