Unroll transcendentals benchmarks by 10x
This reduces the overhead of the benchmark framework and calling the
routine. For example the results for rr::Cos vs. sw::Cos went from
9.6ns / 75ns to 75ns / 725ns. It illustrates that the timings did not
increase tenfold, and the gap between the two widened so it provides
a better measure of the achieved speedup as we further optimize these
transcendental SPIR-V instructions.

This change also copies the full result to an output buffer, to
ensure LLVM doesn't just compute the first element.

Bug: b/158231104
Change-Id: Ia9cde41b32559b16322d074767e986a68f00706d