Support VK_EXT_texel_buffer_alignment promoted features
According to the Vulkan spec:
"This extension adds more expressive alignment requirements for
 uniform and storage texel buffers. Some implementations have
 single texel alignment requirements that cannot be expressed
 via VkPhysicalDeviceLimits::minTexelBufferOffsetAlignment."

For SwiftShader's purposes, we implement support for
VK_EXT_texel_buffer_alignment with the same information
than what's already provided by VkPhysicalDeviceLimits's
minTexelBufferOffsetAlignment.

Note that Vulkan 1.3 doesn't promote the entire extension:
"Functionality in this extension is included in core Vulkan 1.3,
 with the EXT suffix omitted. However, only the properties
 structure is promoted. The feature structure is not promoted.
 The original type name is still available as an alias of the core
 functionality."
so this CL only implements the promoted portion of this extension.

Bug: b/204499872
Change-Id: I378096860387739b50e86fc9ab96e834838e6cb3