Remove unused sin/cos approximations
As detailed in docs/Sin-Cos-Optimization.pdf, it is possible to achieve
Vulkan's `RelaxedPrecision` requirements for sin and cos using only
3 + 1 multiplications, while sine() uses 5 multiplications, which is the
same number as the new Sin() but producing less accurate results. So
sine() and its companions have been removed.

The implementations based on the paper by Mendenhall have also been
removed since they are expensive. They are accurate on the order of
10^-7 which is nearly the full single-precision and far exceeds anything
Vulkan requires. Even if we do need higher precision one day, it should
be tuned to the requirements of that specific case. Most notably the
division should be avoided if possible ([1] looks promising). While the
Mendenhall paper does mention a division-free variant, it appears we
mixed up the polynomial coefficients for the algorithm with division,
with those of the division-free algorithm. So this code needs revising
either way if we revived it.

[1] https://forums.developer.nvidia.com/t/sincospif-implementation-with-improved-performance-and-accuracy/44392

Bug: b/169755552
Change-Id: I7a992ad8f95473c30848d2b243766eba879691fc