Fix shuffling of vectors with a different size mask
LLVM specifies that for the `shufflevector` instruction "The result of
the instruction is a vector whose length is the same as the shuffle
mask": https://llvm.org/docs/LangRef.html#shufflevector-instruction.

Reactor's Nucleus::createShuffleVector() function however takes a raw
pointer to integers as the selection mask, and assumes the size of this
array to be the same as the number of components in the source vectors.
This was also how lowerShuffleVector() worked previously, which was
fine for all uses we had in the x86 specific code.

For other architectures the lowerPack() function was assuming the LLVM
behavior. To reconcile this, the lowerShuffleVector() now takes an
llvm::ArrayRef<int> and behaves like LLVM, and createShuffleVector()
explicitly creates an llvm::SmallVector<int> of the desired size.

Bug: b/165000222
Change-Id: I70ff9f1d9c1f085ccc548a4ec706e7d515b638d5